requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bitarray==2.7.3
bleach==6.0.0
boto3==1.27.0
botocore==1.30.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chatbotAI==0.3.1.3
click==8.1.3
colorama==0.4.6
comm==0.1.3
cryptography==40.0.1
cytoolz==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.2
django-cors-headers==3.10.1
django-forest==0.1.1
django-ipware==4.0.2
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
django-rest-knox==4.2.0
django-static==1.5.6
django-suit==0.2.28
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dynaconf==3.1.12
ecdsa==0.18.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-tester==0.8.0b3
eth-typing==3.3.0
eth-utils==2.1.0
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.17.1
flatbuffers==23.5.26
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.3
gast==0.4.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
hexbytes==0.3.0
hvac==1.1.1
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
isoduration==20.11.0
jax==0.4.13
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
keras==2.12.0
knox==0.1.14
libclang==16.0.0
loguru==0.7.0
lru-dict==1.1.8
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.14.1
marshmallow-jsonapi==0.24.0
matplotlib-inline==0.1.6
mistune==3.0.1
ml-dtypes==0.2.0
multidict==6.0.4
mysqlclient==2.1.1
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
oauthlib==3.2.2
oic==1.5.0
opt-einsum==3.3.0
overrides==7.3.1
packaging==23.1
pandas==1.5.0
pandocfilters==1.5.0
parsimonious==0.9.0
parso==0.8.3
phonenumbers==8.13.7
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.39
protobuf==4.22.1
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
Pygments==2.15.1
pyhcl==0.4.4
pyjwkest==1.4.2
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python-json-logger==2.0.7
pytz==2023.2
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
regex==2023.3.23
reportlab==4.0.4
requests==2.28.2
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlp==3.0.0
rsa==4.9
s3transfer==0.6.1
scipy==1.11.1
semantic-version==2.10.0
Send2Trash==1.8.2
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
twilio==8.3.0
typing_extensions==4.5.0
tzdata==2023.2
Unipath==1.1
uri-template==1.3.0
urllib3==1.26.15
validators==0.20.0
wcwidth==0.2.6
web3==6.0.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==10.4
Werkzeug==2.3.6
whitenoise==6.5.0
win32-setctime==1.1.0
wrapt==1.14.1
xlwt==1.3.0
yarl==1.8.2
